MER: [compositor] Release SHM buffers after we've bound the texture.
We have a copy after that so we don't need it. Clients which
are blocked on buffer release get a good benefit from this earlier
release. For this to work, we need to make sure that we don't
access the buffers after it has been disowned, so we store
all the state we need when we set the back buffer.

Conflicts:
	src/compositor/compositor_api/qwaylandsurfaceitem.cpp
	src/compositor/wayland_wrapper/qwlsurface.cpp

This doesn't apply upstream. The code changed too much and it needs to be
rewritten.