Expose surface's client to QML
Introduce QWaylandClient class that wraps wl_client credentials
and has a method to destroy the client.

Add QWaylandClient object as a QWaylandSurface property.

This come in handy for example when the compositor wants to kill the
process of an unresponsive surface or wants to know client credentails.

WaylandClient typedef is now useless thus is replaced by QWaylandClient.

Change-Id: I997ec459af0f19baef67e7420c90937b197b2e1d
Reviewed-by: Giulio Camuffo <giulio.camuffo@jollamobile.com>