Don't release the last received buffer when the scene graph is stopped.
If there is a valid next buffer the surface is still mapped and when
scene graph is resumed that buffer will be made current so the current
buffer can be released and must be before the client can return from
swapBuffers.  If there is no next buffer, then the surface is  unmapped
and if the current buffer is released there will be no buffer
to render when the scene graph resumes.

Change-Id: I672bd467948613f37faff637dbe35b0bee8c6a1d