MER: Allow KEY_MEDIA to be used as headset key.
[keys] Allow KEY_MEDIA to be used as headset key. Contributes JB#33503

Some kernel input drivers send KEY_MEDIA instead of BTN_MISC from 3.5mm
headset jack. When such thing happens one might want to redefine the
button meaning in the keymap to keep the device specific part on lower
levels and keep the application functionality the same without
increasing the complexity there.

Currently even if one defines in kmap keymap, following

  keycode 226 = ToggleCallHangup

which is then converted to qmap wit kmap2qmap and then passed to the
keyboard plugin with:

  -plugin evdevkeyboard:keymap=/usr/share/qt5/keymaps/mykeymap.qmap

After this QEvdevKeyboardHandler handles the keypress properly:

  QEvdevKeyboardHandler::processKeycode:259 - Processing key event:
  keycode=256, modifiers=00 pressed=1, autorepeat=0  |  plain=517,
  withmod=517, size=518

This is not enough though, but it gets stuck to the qtwayland which has
currently hardcoded list of keycodes. To which this patch adds line to
get it actualy passed forward to applications.

On the long term it would be good if qtwayland would not hardcode such
values, but those would be either passed forward from keymap or defined
in separate file so that one could do device specific configurations.

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>